refactor: 승패 판별을 Dealer로 이동
