+++ src/main/java/blackjack/domain/GameResult.java
@@ -14,32 +14,6 @@ public enum GameResult {
         this.text = text;
     }
 
-    public static GameResult playerResultFrom(final Dealer dealer, final Player player) {
-        if (player.isBust()) {
-            return LOSE;
-        }
-        if (player.isBlackjack() && dealer.isBlackjack()) {
-            return DRAW;
-        }
-        if (player.isBlackjack() || dealer.isBust()) {
-            return WIN;
-        }
-        return getGameResultFromOthers(dealer, player);
-    }
-
-    private static GameResult getGameResultFromOthers(final Dealer dealer, final Player player) {
-        int dealerSum = dealer.calculateDenominations();
-        int playerSum = player.calculateDenominations();
-
-        if (dealerSum < playerSum) {
-            return WIN;
-        }
-        if (dealerSum == playerSum) {
-            return DRAW;
-        }
-        return LOSE;
-    }
-
     public GameResult changeStatusOpposite() {
         if (this == WIN) {
             return LOSE;
