+++ src/main/java/blackjack/domain/user/Dealer.java
@@ -1,6 +1,7 @@
 package blackjack.domain.user;
 
 import blackjack.domain.Card;
+import blackjack.domain.GameResult;
 import java.util.List;
 
 public class Dealer extends Participant {
@@ -24,5 +25,31 @@ public class Dealer extends Participant {
     @Override
     public boolean isImpossibleToAdd() {
         return super.calculateDenominations() > DEALER_DISTRIBUTE_CARD_THRESHOLD;
+    }
+
+    public GameResult judgePlayerResult(final Player player) {
+        if (player.isBust()) {
+            return GameResult.LOSE;
+        }
+        if (player.isBlackjack() && super.isBlackjack()) {
+            return GameResult.DRAW;
+        }
+        if (player.isBlackjack() || super.isBust()) {
+            return GameResult.WIN;
+        }
+        return getGameResultFromOthers(player);
+    }
+
+    private GameResult getGameResultFromOthers(final Player player) {
+        int dealerSum = super.calculateDenominations();
+        int playerSum = player.calculateDenominations();
+
+        if (dealerSum < playerSum) {
+            return GameResult.WIN;
+        }
+        if (dealerSum == playerSum) {
+            return GameResult.DRAW;
+        }
+        return GameResult.LOSE;
     }
 }
